+++ examples/next-aws-bedrock/app/anthropic/page.tsx
@@ -13,7 +13,9 @@ export default function Chat() {
     system: 'red',
     user: 'black',
     function: 'blue',
+    tool: 'purple',
     assistant: 'green',
+    data: 'orange',
   };
 
   return (