+++ examples/next-openai/app/assistant/page.tsx
@@ -7,7 +7,9 @@ const roleToColorMap: Record<Message['role'], string> = {
   system: 'red',
   user: 'black',
   function: 'blue',
+  tool: 'purple',
   assistant: 'green',
+  data: 'orange',
 };
 
 export default function Chat() {
@@ -41,7 +43,16 @@ export default function Chat() {
           style={{ color: roleToColorMap[m.role] }}
         >
           <strong>{`${m.role}: `}</strong>
-          {m.content}
+          {m.role !== 'data' && m.content}
+          {m.role === 'data' && (
+            <>
+              {(m.data as any).description}
+              <br />
+              <pre className={'bg-gray-200'}>
+                {JSON.stringify(m.data, null, 2)}
+              </pre>
+            </>
+          )}
           <br />
           <br />
         </div>