+++ examples/next-openai/app/function-calling/page.tsx
@@ -42,7 +42,9 @@ export default function Chat() {
     system: 'red',
     user: 'black',
     function: 'blue',
+    tool: 'purple',
     assistant: 'green',
+    data: 'orange',
   };
 
   return (