+++ examples/nuxt-openai/pages/function-calling/index.vue
@@ -41,7 +41,9 @@ const roleToColorMap: Record<Message['role'], string> = {
   system: 'red',
   user: 'black',
   function: 'blue',
+  tool: 'purple',
   assistant: 'green',
+  data: 'orange',
 };
 </script>