+++ examples/solidstart-openai/src/routes/function-calling/index.tsx
@@ -41,7 +41,9 @@ export default function Chat() {
     system: 'red',
     user: 'black',
     function: 'blue',
+    tool: 'purple',
     assistant: 'green',
+    data: 'orange',
   };
 
   const handleInputChange: JSX.ChangeEventHandlerUnion<