+++ packages/core/react/use-assistant.ts
@@ -116,13 +116,25 @@ export function experimental_useAssistant({
       )) {
         switch (type) {
           case 'assistant_message': {
-            // append message:
             setMessages(messages => [
               ...messages,
               {
                 id: value.id,
                 role: value.role,
                 content: value.content[0].text.value,
+              },
+            ]);
+            break;
+          }
+
+          case 'data_message': {
+            setMessages(messages => [
+              ...messages,
+              {
+                id: value.id ?? '',
+                role: 'data',
+                content: '',
+                data: value.data,
               },
             ]);
             break;