+++ packages/core/shared/stream-parts.ts
@@ -1,4 +1,10 @@
-import { AssistantMessage, FunctionCall, JSONValue } from './types';
+import {
+  AssistantMessage,
+  DataMessage,
+  FunctionCall,
+  JSONValue,
+  ToolCall,
+} from './types';
 import { StreamString } from './utils';
 
 export interface StreamPart<CODE extends string, NAME extends string, TYPE> {
@@ -72,47 +78,50 @@ const errorStreamPart: StreamPart<'3', 'error', string> = {
   },
 };
 
-const assistantMessage: StreamPart<'4', 'assistant_message', AssistantMessage> =
-  {
-    code: '4',
-    name: 'assistant_message',
-    parse: (value: JSONValue) => {
-      if (
-        value == null ||
-        typeof value !== 'object' ||
-        !('id' in value) ||
-        !('role' in value) ||
-        !('content' in value) ||
-        typeof value.id !== 'string' ||
-        typeof value.role !== 'string' ||
-        value.role !== 'assistant' ||
-        !Array.isArray(value.content) ||
-        !value.content.every(
-          item =>
-            item != null &&
-            typeof item === 'object' &&
-            'type' in item &&
-            item.type === 'text' &&
-            'text' in item &&
-            item.text != null &&
-            typeof item.text === 'object' &&
-            'value' in item.text &&
-            typeof item.text.value === 'string',
-        )
-      ) {
-        throw new Error(
-          '"assistant_message" parts expect an object with an "id", "role", and "content" property.',
-        );
-      }
+const assistantMessageStreamPart: StreamPart<
+  '4',
+  'assistant_message',
+  AssistantMessage
+> = {
+  code: '4',
+  name: 'assistant_message',
+  parse: (value: JSONValue) => {
+    if (
+      value == null ||
+      typeof value !== 'object' ||
+      !('id' in value) ||
+      !('role' in value) ||
+      !('content' in value) ||
+      typeof value.id !== 'string' ||
+      typeof value.role !== 'string' ||
+      value.role !== 'assistant' ||
+      !Array.isArray(value.content) ||
+      !value.content.every(
+        item =>
+          item != null &&
+          typeof item === 'object' &&
+          'type' in item &&
+          item.type === 'text' &&
+          'text' in item &&
+          item.text != null &&
+          typeof item.text === 'object' &&
+          'value' in item.text &&
+          typeof item.text.value === 'string',
+      )
+    ) {
+      throw new Error(
+        '"assistant_message" parts expect an object with an "id", "role", and "content" property.',
+      );
+    }
 
-      return {
-        type: 'assistant_message',
-        value: value as AssistantMessage,
-      };
-    },
-  };
+    return {
+      type: 'assistant_message',
+      value: value as AssistantMessage,
+    };
+  },
+};
 
-const assistantControlData: StreamPart<
+const assistantControlDataStreamPart: StreamPart<
   '5',
   'assistant_control_data',
   {
@@ -146,13 +155,81 @@ const assistantControlData: StreamPart<
   },
 };
 
+const dataMessageStreamPart: StreamPart<'6', 'data_message', DataMessage> = {
+  code: '6',
+  name: 'data_message',
+  parse: (value: JSONValue) => {
+    if (
+      value == null ||
+      typeof value !== 'object' ||
+      !('role' in value) ||
+      !('data' in value) ||
+      typeof value.role !== 'string' ||
+      value.role !== 'data'
+    ) {
+      throw new Error(
+        '"data_message" parts expect an object with a "role" and "data" property.',
+      );
+    }
+
+    return {
+      type: 'data_message',
+      value: value as DataMessage,
+    };
+  },
+};
+
+const toolCallStreamPart: StreamPart<
+  '7',
+  'tool_call',
+  { tool_call: ToolCall[] }
+> = {
+  code: '7',
+  name: 'tool_call',
+  parse: (value: JSONValue) => {
+    if (
+      value == null ||
+      typeof value !== 'object' ||
+      !('tool_calls' in value) ||
+      typeof value.tool_calls !== 'object' ||
+      value.tool_calls == null ||
+      !Array.isArray(value.tool_calls) ||
+      value.tool_calls.some(tc => {
+        tc == null ||
+          typeof tc !== 'object' ||
+          !('id' in tc) ||
+          typeof tc.id !== 'string' ||
+          !('type' in tc) ||
+          typeof tc.type !== 'string' ||
+          !('function' in tc) ||
+          tc.function == null ||
+          typeof tc.function !== 'object' ||
+          !('arguments' in tc.function) ||
+          typeof tc.function.name !== 'string' ||
+          typeof tc.function.arguments !== 'string';
+      })
+    ) {
+      throw new Error(
+        '"tool_call" parts expect an object with a ToolCallPayload.',
+      );
+    }
+
+    return {
+      type: 'tool_call',
+      value: value as unknown as { tool_call: ToolCall[] },
+    };
+  },
+};
+
 const streamParts = [
   textStreamPart,
   functionCallStreamPart,
   dataStreamPart,
   errorStreamPart,
-  assistantMessage,
-  assistantControlData,
+  assistantMessageStreamPart,
+  assistantControlDataStreamPart,
+  dataMessageStreamPart,
+  toolCallStreamPart,
 ] as const;
 
 // union type of all stream parts
@@ -161,8 +238,10 @@ type StreamParts =
   | typeof functionCallStreamPart
   | typeof dataStreamPart
   | typeof errorStreamPart
-  | typeof assistantMessage
-  | typeof assistantControlData;
+  | typeof assistantMessageStreamPart
+  | typeof assistantControlDataStreamPart
+  | typeof dataMessageStreamPart
+  | typeof toolCallStreamPart;
 
 /**
  * Maps the type of a stream part to its value type.
@@ -176,16 +255,20 @@ export type StreamPartType =
   | ReturnType<typeof functionCallStreamPart.parse>
   | ReturnType<typeof dataStreamPart.parse>
   | ReturnType<typeof errorStreamPart.parse>
-  | ReturnType<typeof assistantMessage.parse>
-  | ReturnType<typeof assistantControlData.parse>;
+  | ReturnType<typeof assistantMessageStreamPart.parse>
+  | ReturnType<typeof assistantControlDataStreamPart.parse>
+  | ReturnType<typeof dataMessageStreamPart.parse>
+  | ReturnType<typeof toolCallStreamPart.parse>;
 
 export const streamPartsByCode = {
   [textStreamPart.code]: textStreamPart,
   [functionCallStreamPart.code]: functionCallStreamPart,
   [dataStreamPart.code]: dataStreamPart,
   [errorStreamPart.code]: errorStreamPart,
-  [assistantMessage.code]: assistantMessage,
-  [assistantControlData.code]: assistantControlData,
+  [assistantMessageStreamPart.code]: assistantMessageStreamPart,
+  [assistantControlDataStreamPart.code]: assistantControlDataStreamPart,
+  [dataMessageStreamPart.code]: dataMessageStreamPart,
+  [toolCallStreamPart.code]: toolCallStreamPart,
 } as const;
 
 /**
@@ -194,6 +277,7 @@ export const streamPartsByCode = {
  * - 0: Text from the LLM response
  * - 1: (OpenAI) function_call responses
  * - 2: custom JSON added by the user using `Data`
+ * - 6: (OpenAI) tool_call responses
  *
  * Example:
  * ```
@@ -206,6 +290,7 @@ export const streamPartsByCode = {
  * 0:!
  * 2: { "someJson": "value" }
  * 1: {"function_call": {"name": "get_current_weather", "arguments": "{\\n\\"location\\": \\"Charlottesville, Virginia\\",\\n\\"format\\": \\"celsius\\"\\n}"}}
+ * 6: {"tool_call": {"id": "tool_0", "type": "function", "function": {"name": "get_current_weather", "arguments": "{\\n\\"location\\": \\"Charlottesville, Virginia\\",\\n\\"format\\": \\"celsius\\"\\n}"}}}
  *```
  */
 export const StreamStringPrefixes = {
@@ -213,8 +298,10 @@ export const StreamStringPrefixes = {
   [functionCallStreamPart.name]: functionCallStreamPart.code,
   [dataStreamPart.name]: dataStreamPart.code,
   [errorStreamPart.name]: errorStreamPart.code,
-  [assistantMessage.name]: assistantMessage.code,
-  [assistantControlData.name]: assistantControlData.code,
+  [assistantMessageStreamPart.name]: assistantMessageStreamPart.code,
+  [assistantControlDataStreamPart.name]: assistantControlDataStreamPart.code,
+  [dataMessageStreamPart.name]: dataMessageStreamPart.code,
+  [toolCallStreamPart.name]: toolCallStreamPart.code,
 } as const;
 
 export const validCodes = streamParts.map(part => part.code);