+++ packages/core/tsup.config.ts
@@ -7,6 +7,7 @@ export default defineConfig([
     format: ['cjs', 'esm'],
     external: ['react', 'svelte', 'vue'],
     dts: true,
+    sourcemap: true,
   },
   {
     entry: ['prompts/index.ts'],
@@ -14,6 +15,7 @@ export default defineConfig([
     external: ['react', 'svelte', 'vue'],
     outDir: 'prompts/dist',
     dts: true,
+    sourcemap: true,
   },
   // React APIs
   {
@@ -25,6 +27,7 @@ export default defineConfig([
     format: ['cjs', 'esm'],
     external: ['react', 'svelte', 'vue', 'solid-js'],
     dts: true,
+    sourcemap: true,
   },
   {
     entry: ['react/index.server.ts'],
@@ -32,6 +35,7 @@ export default defineConfig([
     format: ['cjs', 'esm'],
     external: ['react', 'svelte', 'vue', 'solid-js'],
     dts: true,
+    sourcemap: true,
   },
   // Svelte APIs
   {
@@ -41,6 +45,7 @@ export default defineConfig([
     format: ['cjs', 'esm'],
     external: ['react', 'svelte', 'vue', 'solid-js'],
     dts: true,
+    sourcemap: true,
     // `sswr` has some issue with `.es.js` that can't be resolved correctly by
     // vite so we have to bundle it here.
     noExternal: ['sswr'],
@@ -53,6 +58,7 @@ export default defineConfig([
     format: ['cjs', 'esm'],
     external: ['react', 'svelte', 'vue', 'solid-js'],
     dts: true,
+    sourcemap: true,
   },
   // Solid APIs
   {
@@ -62,5 +68,6 @@ export default defineConfig([
     format: ['cjs', 'esm'],
     external: ['react', 'svelte', 'vue', 'solid-js'],
     dts: true,
+    sourcemap: true,
   },
 ]);